# Data/products.go
//we are storing th data of products here
package data

import (
	"fmt"
	"regexp"
	"time"

	"github.com/go-playground/validator"
)

//struct tags are used to add our desired tag names, include, exclude or omit key vaue pairs
//we define a Product struct for our products
//	swagger:model
type Product struct {
	//the id of the user
	//
	//required:true
	//min: 1
	ID          int     `json:"id"`
	Name        string  `json:"name" validate:"required"`
	Description string  `json:"description"`
	Price       float32 `json:"price" validate:"gt=0"`
	SKU         string  `json:"sku" validate:"required,sku"`
	CreatedOn   string  `json:"-"`
	UpdatedOn   string  `json:"-"`
	DeletedOn   string  `json:"-"`
}

//slice of products defined as custom type Products
type Products []*Product

func (p *Product) Validate() error {

	validate := validator.New()
	//add a custom validation for our own tag sku
	validate.RegisterValidation("sku", validateSKU)
	return validate.Struct(p)
}

func validateSKU(fl validator.FieldLevel) bool {
	re := regexp.MustCompile(`[a-z]+-[a-z]+-[a-z]+`)
	matches := re.FindAllString(fl.Field().String(), -1)

	if len(matches) != 1 {
		return false
	}
	return true

}

//returns the current products List
func GetProducts() Products {
	return productList
}

func GetProductsByID(id int) (*Product, error) {
	_, pos, err := findProduct(id)

	if err != nil {
		return nil, err
	}

	return productList[pos], nil

}

//function to Add a new product to the Product List
func AddProduct(p *Product) {
	p.ID = len(productList) + 1
	productList = append(productList, p)

}

//update product at given index id, PUT request
func UpdateProduct(id int, p *Product) error {
	_, pos, err := findProduct(id)

	if err != nil {
		return err
	}

	p.ID = id
	productList[pos] = p

	return nil
}

var ErrProductNotFound = fmt.Errorf("Product Not Found")

func findProduct(id int) (*Product, int, error) {
	for i, p := range productList {
		if p.ID == id {
			return p, i, nil
		}
	}
	return nil, -1, ErrProductNotFound
}

func DeleteProduct(id int) error {
	_, pos, err := findProduct(id)

	if err != nil {
		return err
	}

	productList = append(productList[:pos], productList[pos+1:]...)
	return nil
}

//=================================================================================================
var productList = []*Product{
	&Product{
		ID:          1,
		Name:        "Latte",
		Description: "Frothy Milky Coffee",
		Price:       2.45,
		SKU:         "abc123",
		CreatedOn:   time.Now().UTC().String(),
		UpdatedOn:   time.Now().UTC().String(),
	},
	&Product{
		ID:          2,
		Name:        "Espresso",
		Description: "Short and string coffee without milk",
		Price:       1.99,
		SKU:         "xyz123",
		CreatedOn:   time.Now().UTC().String(),
		UpdatedOn:   time.Now().UTC().String(),
	},
}