# swagger.yaml
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: microservice/Handlers
  Product:
    description: |-
      struct tags are used to add our desired tag names, include, exclude or omit key vaue pairs
      we define a Product struct for our products
    properties:
      description:
        type: string
        x-go-name: Description
      id:
        description: the id of the user
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
      price:
        format: float
        type: number
        x-go-name: Price
      sku:
        type: string
        x-go-name: SKU
    required:
    - id
    type: object
    x-go-package: microservice/Data
info: {}
paths:
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: Adds a New Product in the list
      operationId: addNewProduct
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
  /products/{id}:
    delete:
      description: Removes a product from the given list
      operationId: deleteProducts
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - products
    put:
      description: Updates an existing product in the Products list
      operationId: updateProduct
      parameters:
      - description: The id of the products to update from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - products
  /products{id}:
    get:
      description: Returns a single product
      operationId: listSingleProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - products
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  noContent: {}
  productsResponse:
    description: A list of products returs in the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
swagger: "2.0"